:truck: fix chat user
