+++ src/main/java/com/example/codePicasso/domain/auth/service/AuthService.java
@@ -2,10 +2,10 @@ package com.example.codePicasso.domain.auth.service;
 
 import com.example.codePicasso.domain.auth.dto.request.SigninRequest;
 import com.example.codePicasso.domain.auth.dto.response.JwtTokenResponse;
-import com.example.codePicasso.domain.users.entity.Admin;
-import com.example.codePicasso.domain.users.entity.User;
-import com.example.codePicasso.domain.users.service.AdminConnector;
-import com.example.codePicasso.domain.users.service.UserConnector;
+import com.example.codePicasso.domain.user.entity.Admin;
+import com.example.codePicasso.domain.user.entity.User;
+import com.example.codePicasso.domain.user.service.AdminConnector;
+import com.example.codePicasso.domain.user.service.UserConnector;
 import com.example.codePicasso.global.config.JwtUtil;
 import com.example.codePicasso.global.config.PasswordEncoder;
 import com.example.codePicasso.global.exception.base.InvalidRequestException;
