+++ src/main/java/com/example/codePicasso/domain/chat/dto/request/ChatRequest.java
@@ -1,0 +1,14 @@
+package com.example.codePicasso.domain.chat.dto.request;
+
+import com.example.codePicasso.domain.chat.entity.Chat;
+
+public record ChatRequest(String sender , String message) {
+
+    public Chat toEntity(Long userId) {
+        return Chat.builder()
+                .userId(userId)
+                .username(sender)
+                .content(message)
+                .build();
+    }
+}
