+++ src/main/java/com/example/codePicasso/domain/chat/dto/response/ChatListResponse.java
@@ -1,0 +1,9 @@
+package com.example.codePicasso.domain.chat.dto.response;
+
+import lombok.Builder;
+
+import java.util.List;
+
+@Builder
+public record ChatListResponse(List<ChatResponse> chatsResponseList) {
+}
