+++ src/main/java/com/example/codePicasso/domain/chat/repository/ChatConnectorImpl.java
@@ -1,0 +1,25 @@
+package com.example.codePicasso.domain.chat.repository;
+
+import com.example.codePicasso.domain.chat.entity.Chat;
+import com.example.codePicasso.domain.chat.service.ChatConnector;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Component
+@RequiredArgsConstructor
+public class ChatConnectorImpl implements ChatConnector {
+    private final ChatRepository chatsRepository;
+
+
+    @Override
+    public Chat save(Chat chats) {
+        return chatsRepository.save(chats);
+    }
+
+    @Override
+    public List<Chat> findAll() {
+        return chatsRepository.findAll();
+    }
+}
