+++ src/main/java/com/example/codePicasso/domain/chat/service/ChatConnector.java
@@ -1,0 +1,13 @@
+package com.example.codePicasso.domain.chat.service;
+
+import com.example.codePicasso.domain.chat.entity.Chat;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Component
+public interface ChatConnector {
+    Chat save(Chat chats);
+
+    List<Chat> findAll();
+}
