+++ src/main/java/com/example/codePicasso/domain/chat/service/ChatService.java
@@ -1,0 +1,31 @@
+package com.example.codePicasso.domain.chat.service;
+
+import com.example.codePicasso.domain.chat.dto.request.ChatRequest;
+import com.example.codePicasso.domain.chat.dto.response.ChatResponse;
+import com.example.codePicasso.domain.chat.entity.Chat;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class ChatService {
+
+    private final ChatConnector chatsConnector;
+
+    @Transactional
+    public ChatResponse addMessage(ChatRequest chatsRequest, Long userId) {
+        Chat chats = chatsRequest.toEntity(userId);
+        chatsConnector.save(chats);
+        return chats.toDto();
+    }
+
+    public List<ChatResponse> getChatsHistory() {
+        List<Chat> chats = chatsConnector.findAll();
+        return chats.stream().map(Chat::toDto).toList();
+    }
+}
