--- src/main/java/com/example/codePicasso/domain/chats/dto/response/ChatsListResponse.java
@@ -1,9 +1,0 @@
-package com.example.codePicasso.domain.chats.dto.response;
-
-import lombok.Builder;
-
-import java.util.List;
-
-@Builder
-public record ChatsListResponse(List<ChatsResponse> chatsResponseList) {
-}
