--- src/main/java/com/example/codePicasso/domain/chats/repository/ChatsRepository.java
@@ -1,7 +1,0 @@
-package com.example.codePicasso.domain.chats.repository;
-
-import com.example.codePicasso.domain.chats.entity.Chats;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ChatsRepository extends JpaRepository<Chats, Long> {
-}
