+++ src/main/java/com/example/codePicasso/domain/communities/entity/Community.java
@@ -1,6 +1,6 @@
 package com.example.codePicasso.domain.communities.entity;
 
-import com.example.codePicasso.domain.users.entity.User;
+import com.example.codePicasso.domain.user.entity.User;
 import com.example.codePicasso.global.common.TimeStamp;
 import jakarta.persistence.*;
 import lombok.Getter;
