+++ src/main/java/com/example/codePicasso/domain/exchanges/entity/ItemExchange.java
@@ -1,15 +1,11 @@
 package com.example.codePicasso.domain.exchanges.entity;
 
 import com.example.codePicasso.domain.games.entity.Games;
-import com.example.codePicasso.domain.users.entity.User;
+import com.example.codePicasso.domain.user.entity.User;
 import com.example.codePicasso.global.common.TimeStamp;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Getter
