+++ src/main/java/com/example/codePicasso/domain/games/entity/Games.java
@@ -1,6 +1,6 @@
 package com.example.codePicasso.domain.games.entity;
 
-import com.example.codePicasso.domain.users.entity.Admin;
+import com.example.codePicasso.domain.user.entity.Admin;
 import com.example.codePicasso.global.common.TimeStamp;
 import jakarta.persistence.*;
 import lombok.Getter;
