+++ src/main/java/com/example/codePicasso/domain/user/controller/UseController.java
@@ -1,0 +1,10 @@
+package com.example.codePicasso.domain.user.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/users")
+public class UseController {
+}
