+++ src/main/java/com/example/codePicasso/domain/user/dto/response/AdminResponse.java
@@ -1,0 +1,7 @@
+package com.example.codePicasso.domain.user.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record AdminResponse(String loginId) {
+}
