+++ src/main/java/com/example/codePicasso/domain/user/dto/response/UserResponse.java
@@ -1,0 +1,7 @@
+package com.example.codePicasso.domain.user.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record UserResponse(String loginId, String nickname) {
+}
