+++ src/main/java/com/example/codePicasso/domain/user/entity/UserStatus.java
@@ -1,0 +1,12 @@
+package com.example.codePicasso.domain.user.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum UserStatus {
+    ADMIN("ROLE_ADMIN"),USER("ROLE_USER");
+
+    private final String roles;
+}
