+++ src/main/java/com/example/codePicasso/domain/user/repository/AdminRepository.java
@@ -1,0 +1,11 @@
+package com.example.codePicasso.domain.user.repository;
+
+import com.example.codePicasso.domain.user.entity.Admin;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AdminRepository extends JpaRepository<Admin, Long> {
+
+    boolean existsByLoginId(String s);
+
+    Admin findByLoginId(String s);
+}
