+++ src/main/java/com/example/codePicasso/domain/user/service/UserConnector.java
@@ -1,0 +1,15 @@
+package com.example.codePicasso.domain.user.service;
+
+import com.example.codePicasso.domain.user.entity.User;
+import org.springframework.stereotype.Component;
+
+@Component
+public interface UserConnector {
+    User save(User user);
+
+    boolean existsByLoginId(String s);
+
+    User findByLoginId(String s);
+
+    User findById(Long userId);
+}
