--- src/main/java/com/example/codePicasso/domain/users/dto/request/AdminRequest.java
@@ -1,14 +1,0 @@
-package com.example.codePicasso.domain.users.dto.request;
-
-
-import com.example.codePicasso.domain.users.entity.Admin;
-
-public record AdminRequest(String loginId, String password) {
-
-    public Admin toEntity(String encodedPassword) {
-        return Admin.builder()
-                .loginId(loginId)
-                .password(encodedPassword)
-                .build();
-    }
-}
