--- src/main/java/com/example/codePicasso/domain/users/dto/response/AdminResponse.java
@@ -1,7 +1,0 @@
-package com.example.codePicasso.domain.users.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record AdminResponse(String loginId) {
-}
