--- src/main/java/com/example/codePicasso/domain/users/dto/response/UserResponse.java
@@ -1,8 +1,0 @@
-package com.example.codePicasso.domain.users.dto.response;
-
-import com.example.codePicasso.domain.users.entity.User;
-import lombok.Builder;
-
-@Builder
-public record UserResponse(String loginId, String nickname) {
-}
