--- src/main/java/com/example/codePicasso/domain/users/service/AdminConnector.java
@@ -1,13 +1,0 @@
-package com.example.codePicasso.domain.users.service;
-
-import com.example.codePicasso.domain.users.entity.Admin;
-import org.springframework.stereotype.Component;
-
-@Component
-public interface AdminConnector {
-    boolean existsByLoginId(String s);
-
-    Admin save(Admin admin);
-
-    Admin findByLoginId(String s);
-}
